:recycle: 카드 컴포넌트 스타일 코드 파일 분리
